Refactor useDataProvider for better readability
