+++ packages/ra-core/src/dataProvider/performQuery/index.ts
@@ -1,0 +1,8 @@
+import { doQuery } from './doQuery';
+import {
+    stackCall,
+    stackOptimisticCall,
+    getRemainingStackedCalls,
+} from './stackedCalls';
+
+export { doQuery, stackCall, stackOptimisticCall, getRemainingStackedCalls };
